Add files via upload
Variables such as density, temperature or number of C-atoms, per H-atom, condensed into grains, are plotted as a function of distance from the center of mass of the binary stars. The color of particles indicates the time of dust formation in the following way: black particles have dust formation between 0 and 3 years after the start of the simulation, red particles between 3 and 6 years, green particles between 6 and 9 years, blue particles between 9 and 12 years, magenta particles between 12 and 15 years, light blue particles between 15 and 18 years and orange particles between 18 and 20 years.